Template functions for the tree construction.
