--- packages/Kokkos/src/details/DTK_DetailsTreeConstruction.cpp
@@ -1,272 +1,0 @@
-#include <details/DTK_DetailsAlgorithms.hpp>
-#include <details/DTK_DetailsTreeConstruction.hpp>
-
-#include <algorithm>
-#include <atomic>
-#include <cassert> // TODO: probably want to use something else but fine for now
-
-namespace DataTransferKit
-{
-namespace Details
-{
-namespace Functor
-{
-AssignMortonCodes::AssignMortonCodes( Box const *bounding_boxes,
-                                      unsigned int *morton_codes,
-                                      Box const &scene_bounding_box )
-    : _bounding_boxes( bounding_boxes )
-    , _morton_codes( morton_codes )
-    , _scene_bounding_box( scene_bounding_box )
-{
-}
-
-void AssignMortonCodes::operator()( int const i ) const
-{
-    Point xyz;
-    double a, b;
-    centroid( _bounding_boxes[i], xyz );
-    // scale coordinates with respect to bounding box of the scene
-    for ( int d = 0; d < 3; ++d )
-    {
-        a = _scene_bounding_box[2 * d];
-        b = _scene_bounding_box[2 * d + 1];
-        xyz[d] = ( xyz[d] - a ) / ( b - a );
-    }
-    _morton_codes[i] = morton3D( xyz[0], xyz[1], xyz[2] );
-}
-
-GenerateHierarchy::GenerateHierarchy( unsigned int *sorted_morton_codes,
-                                      Node *leaf_nodes, Node *internal_nodes,
-                                      int n )
-    : _sorted_morton_codes( sorted_morton_codes )
-    , _leaf_nodes( leaf_nodes )
-    , _internal_nodes( internal_nodes )
-    , _n( n )
-{
-}
-
-// from "Thinking Parallel, Part III: Tree Construction on the GPU" by Karras
-void GenerateHierarchy::operator()( int const i ) const
-{
-    // Construct internal nodes.
-    // Find out which range of objects the node corresponds to.
-    // (This is where the magic happens!)
-
-    auto range = determineRange( _sorted_morton_codes, _n, i );
-    int first = range.first;
-    int last = range.second;
-
-    // Determine where to split the range.
-
-    int split = findSplit( _sorted_morton_codes, first, last );
-
-    // Select childA.
-
-    Node *childA;
-    if ( split == first )
-        childA = &_leaf_nodes[split];
-    else
-        childA = &_internal_nodes[split];
-
-    // Select childB.
-
-    Node *childB;
-    if ( split + 1 == last )
-        childB = &_leaf_nodes[split + 1];
-    else
-        childB = &_internal_nodes[split + 1];
-
-    // Record parent-child relationships.
-
-    _internal_nodes[i].children.first = childA;
-    _internal_nodes[i].children.second = childB;
-    childA->parent = &_internal_nodes[i];
-    childB->parent = &_internal_nodes[i];
-}
-
-CalculateBoundingBoxes::CalculateBoundingBoxes(
-    Node const *leaf_nodes, Node *root,
-    std::vector<std::atomic_flag> &atomic_flags )
-    : _leaf_nodes( leaf_nodes )
-    , _root( root )
-    , _atomic_flags( atomic_flags )
-{
-}
-
-void CalculateBoundingBoxes::operator()( int const i ) const
-{
-    Node *node = _leaf_nodes[i].parent;
-    while ( node != _root )
-    {
-        if ( !_atomic_flags[node - _root].test_and_set() )
-            break;
-        for ( Node *child : {node->children.first, node->children.second} )
-            expand( node->bounding_box, child->bounding_box );
-        node = node->parent;
-    }
-    // NOTE: could stop at node != root and then just check that what we
-    // computed earlier (bounding box of the scene) is indeed the union of
-    // the two children.
-}
-}
-
-unsigned int expandBits( unsigned int v )
-{
-    v = ( v * 0x00010001u ) & 0xFF0000FFu;
-    v = ( v * 0x00000101u ) & 0x0F00F00Fu;
-    v = ( v * 0x00000011u ) & 0xC30C30C3u;
-    v = ( v * 0x00000005u ) & 0x49249249u;
-    return v;
-}
-
-// Calculates a 30-bit Morton code for the
-// given 3D point located within the unit cube [0,1].
-unsigned int morton3D( double x, double y, double z )
-{
-    using std::min;
-    using std::max;
-    x = min( max( x * 1024.0, 0.0 ), 1023.0 );
-    y = min( max( y * 1024.0, 0.0 ), 1023.0 );
-    z = min( max( z * 1024.0, 0.0 ), 1023.0 );
-    unsigned int xx = expandBits( (unsigned int)x );
-    unsigned int yy = expandBits( (unsigned int)y );
-    unsigned int zz = expandBits( (unsigned int)z );
-    return xx * 4 + yy * 2 + zz;
-}
-
-// TODO: this is a mess
-// we need a default impl
-#define __clz( x ) __builtin_clz( x )
-// default implementation if nothing else is available
-// Taken from:
-// http://stackoverflow.com/questions/23856596/counting-leading-zeros-in-a-32-bit-unsigned-integer-with-best-algorithm-in-c-pro
-// WARNING: this implementation does **not** support __clz(0) (result should be
-// 32 but this function returns 0)
-int clz( uint32_t x )
-{
-    static const char debruijn32[32] = {
-        0, 31, 9, 30, 3, 8,  13, 29, 2,  5,  7,  21, 12, 24, 28, 19,
-        1, 10, 4, 14, 6, 22, 25, 20, 11, 15, 23, 26, 16, 27, 17, 18};
-    x |= x >> 1;
-    x |= x >> 2;
-    x |= x >> 4;
-    x |= x >> 8;
-    x |= x >> 16;
-    x++;
-    return debruijn32[x * 0x076be629 >> 27];
-}
-
-// TODO: use preprocessor directive to select an implementation
-// it turns out NVDIA's implementation of int __clz(unsigned int x) is
-// slightly different than GCC __builtin_clz
-// this caused a bug in an early implementation of the function that compute
-// the common prefixes between two keys (NB: when i == j)
-int countLeadingZeros( unsigned int x )
-{
-#if defined __CUDACC__
-    // intrinsic function that is only supported in device code
-    // COMMENT: not sure how I am supposed to use it then...
-    return __clz( x );
-
-#elif defined __GNUC__
-    // int __builtin_clz(unsigned int x) result is undefined if x is 0
-    return x != 0 ? __builtin_clz( x ) : 32;
-#else
-    // similar problem with the default implementation
-    return x != 0 ? clz( x ) : 32;
-#endif
-}
-
-int commonPrefix( unsigned int const *k, int n, int i, int j )
-{
-    if ( j < 0 || j > n - 1 )
-        return -1;
-    // our construction algorithm relies on keys being unique so we handle
-    // explicitly case of duplicate Morton codes by augmenting each key by a bit
-    // representation of its index.
-    if ( k[i] == k[j] )
-    {
-        // countLeadingZeros( k[i] ^ k[j] ) == 32
-        return 32 + countLeadingZeros( i ^ j );
-    }
-    return countLeadingZeros( k[i] ^ k[j] );
-}
-
-// from "Thinking Parallel, Part III: Tree Construction on the GPU" by Karras
-int findSplit( unsigned int *sorted_morton_codes, int first, int last )
-{
-    // Identical Morton codes => split the range in the middle.
-
-    unsigned int first_code = sorted_morton_codes[first];
-    unsigned int last_code = sorted_morton_codes[last];
-
-    if ( first_code == last_code )
-        return ( first + last ) >> 1;
-
-    // Calculate the number of highest bits that are the same
-    // for all objects, using the count-leading-zeros intrinsic.
-
-    int common_prefix = __clz( first_code ^ last_code );
-
-    // Use binary search to find where the next bit differs.
-    // Specifically, we are looking for the highest object that
-    // shares more than commonPrefix bits with the first one.
-
-    int split = first; // initial guess
-    int step = last - first;
-
-    do
-    {
-        step = ( step + 1 ) >> 1;     // exponential decrease
-        int new_split = split + step; // proposed new position
-
-        if ( new_split < last )
-        {
-            unsigned int split_code = sorted_morton_codes[new_split];
-            int split_prefix = __clz( first_code ^ split_code );
-            if ( split_prefix > common_prefix )
-                split = new_split; // accept proposal
-        }
-    } while ( step > 1 );
-
-    return split;
-}
-
-// branchless sign function
-// QUESTION: do we want to add it to DTK helpers?
-inline int sgn( int x ) { return ( x > 0 ) - ( x < 0 ); }
-
-Kokkos::pair<int, int> determineRange( unsigned int *sorted_morton_codes, int n,
-                                       int i )
-{
-    using std::min;
-    using std::max;
-    // determine direction of the range (+1 or -1)
-    int direction = sgn( commonPrefix( sorted_morton_codes, n, i, i + 1 ) -
-                         commonPrefix( sorted_morton_codes, n, i, i - 1 ) );
-    assert( direction == +1 || direction == -1 );
-    // compute upper bound for the length of the range
-    int max_step = 2;
-    int common_prefix =
-        commonPrefix( sorted_morton_codes, n, i, i - direction );
-    // compute upper bound for the length of the range
-    while ( commonPrefix( sorted_morton_codes, n, i,
-                          i + direction * max_step ) > common_prefix )
-    {
-        max_step = max_step << 1;
-    }
-    // find the other end using binary search
-    int split = 0;
-    int step = max_step;
-    do
-    {
-        step = step >> 1;
-        if ( commonPrefix( sorted_morton_codes, n, i,
-                           i + ( split + step ) * direction ) > common_prefix )
-            split += step;
-    } while ( step > 1 );
-    int j = i + split * direction;
-    return {min( i, j ), max( i, j )};
-}
-} // end namespace Details
-} // end namespace DataTransferKit
